Complete stage 5.4 и 5.5

+++ JavaService/scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/web/client/StackOverflowClient.java
@@ -4,6 +4,6 @@ import reactor.core.publisher.Mono;
 import ru.tinkoff.edu.java.scrapper.dto.StackOverflowQuestionResponse;
 
 public interface StackOverflowClient {
-    Mono<StackOverflowQuestionResponse> fetchInfoQuestion(int numberQuestion);
+    Mono<StackOverflowQuestionResponse> fetchInfoQuestion(Long numberQuestion);
 }
